# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediBook Pro - Clinic & Hospital Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/firebase:
    post:
      operationId: firebaseAuth
      tags: [auth]
      summary: Authenticate via Firebase (Google/Phone OTP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirebaseAuthBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # USERS
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments (filtered by user or clinic)
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: clinicId
          in: query
          schema:
            type: ["integer", "null"]
        - name: doctorId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Appointments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"

    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "409":
          description: Time slot already booked
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found

    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update/reschedule appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/cancel:
    post:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/complete:
    post:
      operationId: completeAppointment
      tags: [appointments]
      summary: Mark appointment as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /clinics/{clinicId}/queue:
    get:
      operationId: getClinicQueue
      tags: [appointments]
      summary: Get live token queue for a clinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueItem"

  # REVIEWS
...
components:
  schemas:
...
    RegisterUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [patient, clinic_owner, admin]
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    FirebaseAuthBody:
      type: object
      properties:
        idToken:
          type: string
        provider:
          type: string
          enum: [google, phone]
      required:
        - idToken
        - provider
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # USER SCHEMAS
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        firebaseUid:
          type: ["string", "null"]
        isVerified:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isVerified
        - createdAt
        - updatedAt
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        clinicId:
          type: integer
        doctorId:
          type: integer
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        tokenNumber:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        consultationFee:
          type: number
        patientName:
          type: ["string", "null"]
        patientPhone:
          type: ["string", "null"]
        clinicName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - clinicId
        - doctorId
        - appointmentDate
        - appointmentTime
        - tokenNumber
        - status
        - consultationFee
        - createdAt
        - updatedAt
    CreateAppointmentBody:
      type: object
      properties:
        userId:
          type: integer
        clinicId:
          type: integer
        doctorId:
          type: integer
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        notes:
          type: ["string", "null"]
        patientName:
          type: ["string", "null"]
        patientPhone:
          type: ["string", "null"]
      required:
        - userId
        - clinicId
        - doctorId
        - appointmentDate
        - appointmentTime
    UpdateAppointmentBody:
      type: object
      properties:
        appointmentDate:
          type: ["string", "null"]
        appointmentTime:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
    QueueItem:
      type: object
      properties:
        tokenNumber:
          type: integer
        patientName:
          type: ["string", "null"]
        appointmentTime:
          type: string
        status:
          type: string
        doctorName:
          type: ["string", "null"]
        waitingAheadCount:
          type: integer
      required:
        - tokenNumber
        - appointmentTime
        - status
        - waitingAheadCount

    # REVIEW SCHEMAS
...